refactor: reorganize modules and split files
