+++ cli/src/commands.rs
@@ -1,0 +1,5 @@
+mod init;
+mod serve;
+
+pub use init::init;
+pub use serve::serve;
