+++ cli/src/render/mod.rs
@@ -1,12 +1,5 @@
 pub mod typst;
-use ::typst::ecow::EcoString;
-use reflexo_typst::ImmutStr;
 
 pub use self::typst::*;
 pub mod search;
 pub use self::search::*;
-
-pub struct ChapterItem {
-    pub title: EcoString,
-    pub path: Option<ImmutStr>,
-}
